parse to int64

+++ handle/auth.go
@@ -61,7 +61,7 @@ func GenerateTK(c echo.Context, user auth.User) (err error) {
 		"access_token":  accessToken,
 		"refresh_token": refreshToken,
 		"token_type":    "bearer",
-		"expires_in":    accessTokenDuration.Seconds(),
+		"expires_in":    int64(accessTokenDuration.Seconds()),
 		"scope":         user.Scope,
 		"role":          user.Role,
 		"name":          user.Name,
